Provide statistics of gets and contention
Two new attributes are added to the State object for providing to the
user information about the number of gets and which ones suffered from
contention.

By doing the following operation `gets_with_contention/gets` the user can understand
if the connection pool is underprovisioned or if the min idle connections is not properly
configured.